+++ workshop/dm_sentinel/meta.cpp
@@ -1,4 +1,0 @@
-protocol = 1;
-publishedid = 0;
-name = "Sentinel Deathmatch - Sentinel Bridge (Server)";
-timestamp = 0;